pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "satmas"
version = "0.1.0"
description = "Multi-Agent Resource Allocation using PySAT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-sat>=1.8.dev16",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
